feat: Add GLWB pricer, examples, and data generation scripts
## New Features
- Integrate GLWBPricer into ProductRegistry with lazy initialization
- Add 4 production examples (fair rates, competitive positioning, GLWB, stress testing)
- Add data generation utilities (payoff truth tables, synthetic rates, GLWB golden vectors)

## Documentation
- Add stress testing guide (docs/stress_testing/)
- Update getting started and pricing guides
- Add CONTRIBUTING.md and SECURITY.md governance docs

## Improvements
- Lazy-initialize GLWB pricer to support negative rate environments
- Relax mypy strictness for prototype modules (regulatory, stress testing, GLWB)
- Update public API exports in __init__.py

## Tests
- Add GLWB registry integration tests
- Add golden file validation tests
- Add reference data files for validation
- All 2697 tests pass, 85% coverage

+++ src/annuity_pricing/options/payoffs/fia.py
@@ -543,7 +543,7 @@ class MonthlyAveragePayoff(BasePayoff):
 
         # Calculate average of monthly index levels
         avg_value = np.mean(path.values)
-        avg_return = (avg_value - path.initial_value) / path.initial_value
+        avg_return = float((avg_value - path.initial_value) / path.initial_value)
 
         return self.calculate(avg_return)
